finish part identity

+++ src/monolith/Infrastructure/ImplementationContract/Repositories/NpgsqlParameters/ProductParameter.cs
@@ -10,7 +10,7 @@ public static class ProductParameter
         command.Parameters.AddWithValue("@Price", entity.Price);
         command.Parameters.AddWithValue("@StockQuantity", entity.StockQuantity);
         command.Parameters.AddWithValue("@ImageUrl", entity.ImageUrl.ToDbValue());
-        command.Parameters.AddWithValue("@Status", entity.Status);
+        command.Parameters.AddWithValue("@Status", (int)entity.Status);
         command.Parameters.AddWithValue("@CreatedAt", entity.CreatedAt);
         command.Parameters.AddWithValue("@CreatedBy", entity.CreatedBy.ToDbValue());
         command.Parameters.AddWithValue("@CreatedByIp", entity.CreatedByIp.ToDbValue());
@@ -24,7 +24,7 @@ public static class ProductParameter
         command.Parameters.AddWithValue("@Price", entity.Price);
         command.Parameters.AddWithValue("@StockQuantity", entity.StockQuantity);
         command.Parameters.AddWithValue("@ImageUrl", entity.ImageUrl.ToDbValue());
-        command.Parameters.AddWithValue("@Status", entity.Status);
+        command.Parameters.AddWithValue("@Status", (int)entity.Status);
         command.Parameters.AddWithValue("@UpdatedAt", entity.UpdatedAt.ToDbValue());
         command.Parameters.AddWithValue("@UpdatedBy", entity.UpdatedBy.ToDbValue());
         command.Parameters.AddWithValue("@UpdatedByIp", entity.UpdatedByIp.ToDbValue());
