+++ src/monolith/Infrastructure/ImplementationContract/Repositories/NpgsqlParameters/RoleParameter.cs
@@ -8,7 +8,7 @@ public static class RoleParameter
         command.Parameters.AddWithValue("@Name", entity.Name);
         command.Parameters.AddWithValue("@RoleKey", entity.RoleKey);
         command.Parameters.AddWithValue("@Description", entity.Description.ToDbValue());
-        command.Parameters.AddWithValue("@Status", entity.Status);
+        command.Parameters.AddWithValue("@Status",(int)entity.Status);
         command.Parameters.AddWithValue("@CreatedAt", entity.CreatedAt);
         command.Parameters.AddWithValue("@CreatedBy", entity.CreatedBy.ToDbValue());
         command.Parameters.AddWithValue("@CreatedByIp", entity.CreatedByIp.ToDbValue());
@@ -20,7 +20,7 @@ public static class RoleParameter
         command.Parameters.AddWithValue("@Name", entity.Name);
         command.Parameters.AddWithValue("@RoleKey", entity.RoleKey);
         command.Parameters.AddWithValue("@Description", entity.Description.ToDbValue());
-        command.Parameters.AddWithValue("@Status", entity.Status);
+        command.Parameters.AddWithValue("@Status", (int)entity.Status);
         command.Parameters.AddWithValue("@UpdatedAt", entity.UpdatedAt.ToDbValue());
         command.Parameters.AddWithValue("@UpdatedBy", entity.UpdatedBy.ToDbValue());
         command.Parameters.AddWithValue("@UpdatedByIp", entity.UpdatedByIp.ToDbValue());
