+++ src/monolith/Infrastructure/ImplementationContract/Repositories/NpgsqlParameters/UserParameter.cs
@@ -25,7 +25,7 @@ public static class UserParameter
         command.Parameters.AddWithValue("@PhoneNumber", entity.PhoneNumber);
         command.Parameters.AddWithValue("@UserName", entity.UserName);
         command.Parameters.AddWithValue("@PasswordHash", entity.PasswordHash);
-        command.Parameters.AddWithValue("@Status", entity.Status);
+        command.Parameters.AddWithValue("@Status",(int)entity.Status);
         command.Parameters.AddWithValue("@UpdatedAt", entity.UpdatedAt.ToDbValue());
         command.Parameters.AddWithValue("@UpdatedBy", entity.UpdatedBy.ToDbValue());
         command.Parameters.AddWithValue("@UpdatedByIp", entity.UpdatedByIp.ToDbValue());
