+++ src/monolith/Infrastructure/ImplementationContract/Repositories/NpgsqlParameters/UserRoleParameter.cs
@@ -7,7 +7,7 @@ public static class UserRoleParameter
         command.Parameters.AddWithValue("@Id", entity.Id);
         command.Parameters.AddWithValue("@UserId", entity.UserId);
         command.Parameters.AddWithValue("@RoleId", entity.RoleId);
-        command.Parameters.AddWithValue("@Status", entity.Status);
+        command.Parameters.AddWithValue("@Status", (int)entity.Status);
         command.Parameters.AddWithValue("@CreatedAt", entity.CreatedAt);
         command.Parameters.AddWithValue("@CreatedBy", entity.CreatedBy.ToDbValue());
         command.Parameters.AddWithValue("@CreatedByIp", entity.CreatedByIp.ToDbValue());
@@ -18,7 +18,7 @@ public static class UserRoleParameter
         command.Parameters.AddWithValue("@Id", entity.Id);
         command.Parameters.AddWithValue("@UserId", entity.UserId);
         command.Parameters.AddWithValue("@RoleId", entity.RoleId);
-        command.Parameters.AddWithValue("@Status", entity.Status);
+        command.Parameters.AddWithValue("@Status", (int)entity.Status);
         command.Parameters.AddWithValue("@UpdatedAt", entity.UpdatedAt.ToDbValue());
         command.Parameters.AddWithValue("@UpdatedBy", entity.UpdatedBy.ToDbValue());
         command.Parameters.AddWithValue("@UpdatedByIp", entity.UpdatedByIp.ToDbValue());
